move old yaml parser to deprecated
