+++ ydb/core/driver_lib/cli_utils/cli_cmds_config.cpp
@@ -1,6 +1,6 @@
 #include <ydb/core/protos/blobstorage.pb.h>
 #include <ydb/core/protos/config.pb.h>
-#include <ydb/library/yaml_config/yaml_config_parser.h>
+#include <ydb/library/yaml_config/deprecated/yaml_config_parser.h>
 #include "cli.h"
 #include "cli_cmds.h"
 #include "proto_common.h"
