+++ ydb/core/driver_lib/cli_utils/cli_cmds_server.cpp
@@ -4,7 +4,7 @@
 #include <ydb/core/base/path.h>
 #include <ydb/library/yaml_config/yaml_config.h>
 #include <ydb/core/driver_lib/run/run.h>
-#include <ydb/library/yaml_config/yaml_config_parser.h>
+#include <ydb/library/yaml_config/deprecated/yaml_config_parser.h>
 #include <ydb/public/lib/deprecated/kicli/kicli.h>
 #include <util/digest/city.h>
 #include <util/random/random.h>
