+++ ydb/library/yaml_config/deprecated/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    yaml_config_parser.cpp
+    yaml_config_parser.h
+)
+
+PEERDIR(
+    ydb/library/yaml_config
+)
+
+END()
