+++ ydb/library/yaml_config/yaml_config.cpp
@@ -1,6 +1,6 @@
 #include "yaml_config.h"
 
-#include "yaml_config_parser.h"
+#include <ydb/library/yaml_config/deprecated/yaml_config_parser.h>
 
 #include <ydb/core/base/appdata.h>
